+++ app/templates/main/base.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="ru">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="shortcut icon" type="image/png" href="{{url_for('static', filename = 'images/favicon.ico')}}">
+
+        <title>
+            {% block title %} NoSQL {% endblock %}
+        </title>
+
+        {% block extra_css %} {% endblock %}
+
+        {% block extra_javascript %} {% endblock %}
+
+    </head>
+    <body>
+
+        {% block content %} {% endblock %}
+
+    </body>
+</html>
