+++ app/templates/main/index.html
@@ -1,0 +1,5 @@
+{% extends "main/base.html" %}
+
+{% block content %}
+
+{% endblock %}
