+++ config/config.ini
@@ -1,0 +1,4 @@
+[db]
+name = nosql1h19-jira-stats
+host = 127.0.0.1
+port = 27017
