+++ jira-stats.wsgi
@@ -1,0 +1,1 @@
+from manage import app as application
