--- site/app/api/__init__.py
@@ -1,4 +1,0 @@
-from flask import Blueprint
-
-
-api = Blueprint('api', __name__)
