--- site/app/main/views.py
@@ -1,7 +1,0 @@
-from flask import render_template
-from . import main
-
-
-@main.route('/')
-def index():
-    return render_template("main/index.html", greeting='HelloWorld')
