--- site/app/templates/main/index.html
@@ -1,1 +1,0 @@
-{{ greeting }}
