--- site/config.ini
@@ -1,4 +1,0 @@
-[db]
-name = test
-host = 127.0.0.1
-port = 27017
